Eliminated some warnings

+++ WalletsLib/CheckRecipSigner.cpp
@@ -226,7 +226,7 @@ bool CheckRecipSigner::GetInputAddressList(const std::shared_ptr<spdlog::logger>
    const auto &cbOutputTXs = [this, cbTXs, cb](std::vector<Tx> txs) {
       for (const auto &tx : txs) {
          for (size_t i = 0; i < tx.getNumTxIn(); ++i) {
-            TxIn in = tx.getTxInCopy(i);
+            TxIn in = tx.getTxInCopy((int)i);
             OutPoint op = in.getOutPoint();
             txHashSet_.insert(op.getTxHash());
             txOutIdx_[op.getTxHash()].insert(op.getTxOutIndex());
@@ -275,7 +275,7 @@ int TxChecker::receiverIndex(const bs::Address &addr) const
       }
       const auto &txAddr = bs::Address::fromTxOut(out);
       if (addr.id() == txAddr.id()) {
-         return i;
+         return (int)i;
       }
    }
    return -1;
